+++ merlin/main.py
@@ -52,14 +52,7 @@ from merlin import VERSION, router
 from merlin.ascii_art import banner_small
 from merlin.examples.generator import list_examples, setup_example
 from merlin.log_formatter import setup_logging
-from merlin.server.server_setup import (
-    ServerStatus,
-    create_server_config,
-    get_server_status,
-    pull_server_image,
-    start_server,
-    stop_server,
-)
+from merlin.server.server_commands import config_server, init_server, restart_server, start_server, status_server, stop_server
 from merlin.spec.expansion import RESERVED, get_spec_with_expansion
 from merlin.spec.specification import MerlinSpec
 from merlin.study.study import MerlinStudy
@@ -350,30 +343,19 @@ def process_monitor(args):
     LOG.info("Monitor: ... stop condition met")
 
 
-def process_server(args):
+def process_server(args: Namespace):
     if args.commands == "init":
-        if not create_server_config():
-            LOG.info("Merlin server initialization failed.")
-            return
-        if pull_server_image():
-            LOG.info("New merlin server image fetched")
-        LOG.info("Merlin server initialization successful.")
+        init_server()
     elif args.commands == "start":
         start_server()
     elif args.commands == "stop":
         stop_server()
     elif args.commands == "status":
-        current_status = get_server_status()
-        if current_status == ServerStatus.NOT_INITALIZED:
-            LOG.info("Merlin server has not been initialized.")
-            LOG.info("Please initalize server by running 'merlin server init'")
-        elif current_status == ServerStatus.MISSING_CONTAINER:
-            LOG.info("Unable to find server image.")
-            LOG.info("Ensure there is a .sif file in merlin server directory.")
-        elif current_status == ServerStatus.NOT_RUNNING:
-            LOG.info("Merlin server is not running.")
-        elif current_status == ServerStatus.RUNNING:
-            LOG.info("Merlin server is running.")
+        status_server()
+    elif args.commands == "restart":
+        restart_server()
+    elif args.commands == "config":
+        config_server(args)
 
 
 def setup_argparse() -> None:
@@ -591,6 +573,7 @@ def setup_argparse() -> None:
         help="Manage broker and results server for merlin workflow.",
         formatter_class=ArgumentDefaultsHelpFormatter,
     )
+    server.set_defaults(func=process_server)
 
     server_commands: ArgumentParser = server.add_subparsers(dest="commands")
 
@@ -625,6 +608,101 @@ def setup_argparse() -> None:
         formatter_class=ArgumentDefaultsHelpFormatter,
     )
     server_stop.set_defaults(func=process_server)
+
+    server_stop: ArgumentParser = server_commands.add_parser(
+        "restart",
+        help="Restart merlin server instance",
+        description="Restart server.",
+        formatter_class=ArgumentDefaultsHelpFormatter,
+    )
+    server_stop.set_defaults(func=process_server)
+
+    server_config: ArgumentParser = server_commands.add_parser(
+        "config",
+        help="Making configurations for to the merlin server instance.",
+        description="Config server.",
+        formatter_class=ArgumentDefaultsHelpFormatter,
+    )
+    server_config.add_argument(
+        "-ip",
+        "--ipaddress",
+        action="store",
+        type=str,
+        # default="127.0.0.1",
+        help="Set the binded IP address for the merlin server container.",
+    )
+    server_config.add_argument(
+        "-p",
+        "--port",
+        action="store",
+        type=int,
+        # default=6379,
+        help="Set the binded port for the merlin server container.",
+    )
+    server_config.add_argument(
+        "-pwd",
+        "--password",
+        action="store",
+        type=str,
+        # default="~/.merlin/redis.pass",
+        help="Set the password file to be used for merlin server container.",
+    )
+    server_config.add_argument(
+        "--add-user",
+        action="store",
+        nargs=2,
+        type=str,
+        help="Create a new user for merlin server instance. (Provide both username and password)",
+    )
+    server_config.add_argument("--remove-user", action="store", type=str, help="Remove an exisiting user.")
+    server_config.add_argument(
+        "-d",
+        "--directory",
+        action="store",
+        type=str,
+        # default="./",
+        help="Set the working directory of the merlin server container.",
+    )
+    server_config.add_argument(
+        "-ss",
+        "--snapshot-seconds",
+        action="store",
+        type=int,
+        # default=300,
+        help="Set the number of seconds merlin server waits before checking if a snapshot is needed.",
+    )
+    server_config.add_argument(
+        "-sc",
+        "--snapshot-changes",
+        action="store",
+        type=int,
+        # default=100,
+        help="Set the number of changes that are required to be made to the merlin server before a snapshot is made.",
+    )
+    server_config.add_argument(
+        "-sf",
+        "--snapshot-file",
+        action="store",
+        type=str,
+        # default="dump.db",
+        help="Set the snapshot filename for database dumps.",
+    )
+    server_config.add_argument(
+        "-am",
+        "--append-mode",
+        action="store",
+        type=str,
+        # default="everysec",
+        help="The appendonly mode to be set. The avaiable options are always, everysec, no.",
+    )
+    server_config.add_argument(
+        "-af",
+        "--append-file",
+        action="store",
+        type=str,
+        # default="appendonly.aof",
+        help="Set append only filename for merlin server container.",
+    )
 
     return parser